Edit/destroy books authorization

+++ app/controllers/books_controller.rb
@@ -10,6 +10,7 @@ class BooksController < ApplicationController
 
   def create
     @book = Book.new(book_params)
+    @book.reader_id = session[:reader_id]
     if @book.save
       redirect_to books_path, notice: "You added new book into your library"
     else
@@ -29,6 +30,7 @@ class BooksController < ApplicationController
 
   def edit
     @book = Book.find(params[:id])
+    raise AccessDenied unless @book.owned_by?(current_user)
   end
 
   def update
@@ -43,6 +45,7 @@ class BooksController < ApplicationController
 
   def destroy
     book = Book.find(params[:id])
+    raise AccessDenied unless book.owned_by?(current_user)
     book.destroy
     redirect_to books_path
   end
